chore: move utils to utils classes
